refactor: move form inputs into ui library
