+++ apps/web/src/app/admin/events/CreateEventForm.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Fieldset, Form, Input } from '@eventuras/forms';
+import { Fieldset, Form, Input } from '@eventuras/ratio-ui/forms';
 import { Button } from '@eventuras/ratio-ui';
 import { DATA_TEST_ID } from '@eventuras/utils';
 import { useTranslations } from 'next-intl';
