+++ lab/communitas-front/tsconfig.json
@@ -22,10 +22,8 @@
       "@/*": ["./src/*"],
       "@eventuras/datatable": ["../../libs/datatable/src/"],
       "@eventuras/datatable/*": ["../../libs/datatable/src/*"],
-      "@eventuras/forms": ["../../libs/forms/src/"],
-      "@eventuras/forms/*": ["../../libs/forms/src/*"],
       "@eventuras/ratio-ui": ["../../libs/ratio-ui/src/"],
-      "@eventuras/ratio-ui/*": ["../../libs/ratio-ui/src/*"],
+      "@eventuras/ratio-ui/*": ["../../libs/ratio-ui/src/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
