+++ libs/datatable/tsconfig.json
@@ -1,18 +1,16 @@
 {
   "extends": "@eventuras/typescript-config/react-library.json",
-    "compilerOptions": {
+  "compilerOptions": {
     "paths": {
-      "@eventuras/datatable/*":["../../libs/datatable/src/*"],
-      "@eventuras/forms/*":["../../libs/forms/src/*"],
-      "@eventuras/ratio-ui":["../../libs/ratio-ui/src"],
-      "@eventuras/ratio-ui/*":["../../libs/ratio-ui/src/*"],
+      "@eventuras/datatable/*": ["../../libs/datatable/src/*"],
+      "@eventuras/ratio-ui": ["../../libs/ratio-ui/src"],
+      "@eventuras/ratio-ui/*": ["../../libs/ratio-ui/src/*"]
     },
     "plugins": [
       {
         "name": "next"
       }
-    ],
-
+    ]
   },
   "include": ["src"],
   "exclude": ["node_modules", "dist"]
