--- libs/forms/eslint.config.js
@@ -1,4 +1,0 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  extends: ['@eventuras/eslint-config/react-library.js'],
-};
