--- libs/forms/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@eventuras/forms",
-  "version": "0.1.0",
-  "main": "src/index.tsx",
-  "peerDependencies": {
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0"
-  },
-  "devDependencies": {
-    "@eventuras/eslint-config": "*",
-    "@eventuras/typescript-config": "*"
-  }
-}
