--- libs/forms/src/index.tsx
@@ -1,7 +1,0 @@
-export { DebouncedInput } from './inputs/DebouncedInput';
-export { Fieldset } from './common/Fieldset';
-export { Form } from './common/Form';
-export { formStyles } from './styles/formStyles';
-export { InputLabel } from './common/InputLabel';
-export type { InputProps } from './inputs/InputProps.ts';
-export { Input } from './inputs/Input';
