--- libs/forms/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "@eventuras/typescript-config/react-library.json",
-    "compilerOptions": {
-    "paths": {
-      "@/utils/*":["../../apps/web/src/utils/*"],
-      "@eventuras/ratio-ui":["../../libs/ratio-ui/src"],
-      "@eventuras/ratio-ui/*":["../../libs/ratio-ui/src/*"],
-    },
-  },
-  "include": ["src", "../../apps/web/src/components/eventuras/UserLookup.tsx"],
-  "exclude": ["node_modules", "dist"]
-}
