+++ libs/smartform/package.json
@@ -4,7 +4,6 @@
   "version": "0.1.0",
   "dependencies": {
     "@eventuras/ratio-ui": "*",
-    "@eventuras/forms": "*",
     "react-hook-form": "^7.55.0"
   },
   "peerDependencies": {
